requirements.txt
selenium==4.9.0
opencv-python==4.7.0.72
pytesseract==0.3.10
tensorflow==2.12.0
python-dotenv==1.0.0